fix targeted report folder issues + update to bruker json
